% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answers.R
\name{get_answers}
\alias{get_answers}
\title{Retrieve answers}
\usage{
get_answers(answers)
}
\arguments{
\item{answers}{from \code{js$renderSurvey()} internally assigned to \code{id_answers} in \code{render_survey()}. Pass as \code{input$id_answers}.}
}
\value{

}
\description{
Handles the answers passed from the \code{js$renderSurvey()} function (called in \code{render_survey()}).
}
