% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey.R
\name{import_survey}
\alias{import_survey}
\title{Load json file}
\usage{
import_survey(path)
}
\arguments{
\item{path}{Path to json file containing the survey structure and flow.}
}
\value{

}
\description{
Leverage the survey creator here \url{https://surveyjs.io/create-survey}.
}
