% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerDependencies.R
\name{registerDependencies}
\alias{registerDependencies}
\title{Register Dependencies}
\usage{
registerDependencies(version, theme)
}
\arguments{
\item{version}{survey-jquery version.}

\item{theme}{see \code{useSurveyjr()} for a description.}
}
\value{
\code{htmlDependeny} list
}
\description{
Dependencies: package{shinyjs} and SurveyJS framework. Adds resource path to internal js functions.
}
\keyword{internal}
