% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_survey.R
\name{render_survey}
\alias{render_survey}
\title{Render the Survey}
\usage{
render_survey(id_container, id_answers, survey_json)
}
\arguments{
\item{id_container}{id which has to match id from \code{survey_output()}.}

\item{id_answers}{will write answers internally to \code{input$id_answers}.}

\item{survey_json}{json object imported with \code{import_survey()}.}
}
\value{

}
\description{
Main magic happens in \code{js$renderSurvey()} JavaScript function.
}
