% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SURVEY}
\alias{SURVEY}
\title{List of sample surveys}
\format{
Each element has a json data structure
}
\source{
\url{https://surveyjs.io/create-survey}
}
\usage{
SURVEY
}
\description{
import_survey(path_to_json)
}
\keyword{datasets}
