% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_output.R
\name{survey_output}
\alias{survey_output}
\title{Container for rendered Survey}
\usage{
survey_output(id)
}
\arguments{
\item{id}{to communicate with \code{render_survey()}.}
}
\value{

}
\description{
Simple div
}
