% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyApp.R
\name{surveyApp}
\alias{surveyApp}
\title{Example app}
\usage{
surveyApp(
  bootswatch = "darkly",
  survey = c("cancellation", "market_fit", "patient")
)
}
\arguments{
\item{bootswatch}{character; one of bslib::bootswatch_themes().}

\item{survey}{one of names(surveyjR::SURVEY).}
}
\value{

}
\description{
Example app
}
