% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useSurveyjr.R
\name{useSurveyjr}
\alias{useSurveyjr}
\title{Register surveyjR package}
\usage{
useSurveyjr(
  version = "1.9.4",
  theme = c("modern", "default", "orange", "darkblue", "darkrose", "stone", "winter",
    "winterstone", "bootstrap")
)
}
\arguments{
\item{version}{survey-jquery version.}

\item{theme}{if you use a bootswatch theme bootrap needs to be chosen (default).}
}
\value{
registration objects (lists) retrieved from internal functions...
}
\description{
Calls internal functions \code{registerDependencies()} and \code{registerJs()}. Additionally writes
theme to .surveyjR environment such that other functions can retrieve it from there.
}
